relation: Add common arithmetic helpers to `ConstraintSystem` trait
This involves removing the implementation that were previously done
outside of a trait implementation.

We implement inside of a trait so that gadgets may be easily defined
over both the single and multiprover implementations.
