+++ plonk/src/multiprover/proof_system/constraint_system.rs
@@ -18,11 +18,14 @@ use mpc_relation::{
     constants::{compute_coset_representatives, GATE_WIDTH, N_MUL_SELECTORS},
     errors::CircuitError,
     gates::{
-        AdditionGate, BoolGate, ConstantGate, EqualityGate, Gate, IoGate, MultiplicationGate,
-        PaddingGate, SubtractionGate,
+        AdditionGate, BoolGate, ConstantAdditionGate, ConstantGate, ConstantMultiplicationGate,
+        EqualityGate, Gate, IoGate, LinCombGate, MulAddGate, MultiplicationGate, PaddingGate,
+        SubtractionGate,
     },
     ConstraintSystem, GateId, Variable, WireId,
 };
+
+use crate::multiprover::gadgets::{next_multiple, scalar};
 
 // --------------------
 // | Traits and Types |
@@ -738,6 +741,13 @@ impl<C: CurveGroup> ConstraintSystem<C::ScalarField> for MpcPlonkCircuit<C> {
         false
     }
 
+    fn pad_gates(&mut self, n: usize) {
+        let wire_vars = &[self.zero(), self.zero(), 0, 0, 0];
+        for _ in 0..n {
+            self.insert_gate(wire_vars, Box::new(EqualityGate)).unwrap();
+        }
+    }
+
     fn enforce_constant(
         &mut self,
         var: Variable,
@@ -747,6 +757,23 @@ impl<C: CurveGroup> ConstraintSystem<C::ScalarField> for MpcPlonkCircuit<C> {
 
         let wire_vars = &[0, 0, 0, 0, var];
         self.insert_gate(wire_vars, Box::new(ConstantGate(constant)))?;
+        Ok(())
+    }
+
+    fn enforce_bool(&mut self, a: Variable) -> Result<(), CircuitError> {
+        self.check_var_bound(a)?;
+
+        let wire_vars = &[a, a, 0, 0, a];
+        self.insert_gate(wire_vars, Box::new(BoolGate))?;
+        Ok(())
+    }
+
+    fn enforce_equal(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError> {
+        self.check_var_bound(a)?;
+        self.check_var_bound(b)?;
+
+        let wire_vars = &[a, b, 0, 0, 0];
+        self.insert_gate(wire_vars, Box::new(EqualityGate))?;
         Ok(())
     }
 
@@ -813,28 +840,205 @@ impl<C: CurveGroup> ConstraintSystem<C::ScalarField> for MpcPlonkCircuit<C> {
         Ok(c)
     }
 
-    fn enforce_bool(&mut self, a: Variable) -> Result<(), CircuitError> {
-        self.check_var_bound(a)?;
+    fn lc_gate(
+        &mut self,
+        wires: &[Variable; GATE_WIDTH + 1],
+        coeffs: &[C::ScalarField; GATE_WIDTH],
+    ) -> Result<(), CircuitError> {
+        self.check_vars_bound(wires)?;
 
-        let wire_vars = &[a, a, 0, 0, a];
-        self.insert_gate(wire_vars, Box::new(BoolGate))?;
+        let wire_vars = [wires[0], wires[1], wires[2], wires[3], wires[4]];
+        self.insert_gate(&wire_vars, Box::new(LinCombGate { coeffs: *coeffs }))?;
         Ok(())
     }
 
-    fn enforce_equal(&mut self, a: Variable, b: Variable) -> Result<(), CircuitError> {
-        self.check_var_bound(a)?;
-        self.check_var_bound(b)?;
+    fn lc(
+        &mut self,
+        wires_in: &[Variable; GATE_WIDTH],
+        coeffs: &[C::ScalarField; GATE_WIDTH],
+    ) -> Result<Variable, CircuitError> {
+        self.check_vars_bound(wires_in)?;
 
-        let wire_vars = &[a, b, 0, 0, 0];
-        self.insert_gate(wire_vars, Box::new(EqualityGate))?;
+        let vals_in: Vec<AuthenticatedScalarResult<C>> = wires_in
+            .iter()
+            .map(|&var| self.witness(var))
+            .collect::<Result<Vec<_>, CircuitError>>()?;
+
+        // calculate y as the linear combination of coeffs and vals_in
+        let y_val = vals_in
+            .iter()
+            .zip(coeffs.iter())
+            .map(|(val, coeff)| val * scalar!(*coeff))
+            .sum();
+        let y = self.create_variable(y_val)?;
+
+        let wires = [wires_in[0], wires_in[1], wires_in[2], wires_in[3], y];
+        self.lc_gate(&wires, coeffs)?;
+        Ok(y)
+    }
+
+    fn mul_add_gate(
+        &mut self,
+        wires: &[Variable; GATE_WIDTH + 1],
+        q_muls: &[C::ScalarField; N_MUL_SELECTORS],
+    ) -> Result<(), CircuitError> {
+        self.check_vars_bound(wires)?;
+
+        let wire_vars = [wires[0], wires[1], wires[2], wires[3], wires[4]];
+        self.insert_gate(&wire_vars, Box::new(MulAddGate { coeffs: *q_muls }))?;
         Ok(())
     }
 
-    fn pad_gates(&mut self, n: usize) {
-        let wire_vars = &[self.zero(), self.zero(), 0, 0, 0];
-        for _ in 0..n {
-            self.insert_gate(wire_vars, Box::new(EqualityGate)).unwrap();
+    fn mul_add(
+        &mut self,
+        wires_in: &[Variable; GATE_WIDTH],
+        q_muls: &[C::ScalarField; N_MUL_SELECTORS],
+    ) -> Result<Variable, CircuitError> {
+        self.check_vars_bound(wires_in)?;
+
+        let vals_in: Vec<AuthenticatedScalarResult<C>> = wires_in
+            .iter()
+            .map(|&var| self.witness(var))
+            .collect::<Result<Vec<_>, CircuitError>>()?;
+
+        // calculate y as the mul-addition of coeffs and vals_in
+        let y_val = scalar!(q_muls[0]) * &vals_in[0] * &vals_in[1]
+            + scalar!(q_muls[1]) * &vals_in[2] * &vals_in[3];
+        let y = self.create_variable(y_val)?;
+
+        let wires = [wires_in[0], wires_in[1], wires_in[2], wires_in[3], y];
+        self.mul_add_gate(&wires, q_muls)?;
+        Ok(y)
+    }
+
+    /// Create a variable representing the sum of a list of variables
+    ///
+    /// Return an error if variables are invalid
+    fn sum(&mut self, elems: &[Variable]) -> Result<Variable, CircuitError> {
+        if elems.is_empty() {
+            return Err(CircuitError::ParameterError(
+                "Sum over an empty slice of variables is undefined".to_string(),
+            ));
         }
+        self.check_vars_bound(elems)?;
+
+        let mut elem_iter = elems
+            .iter()
+            .map(|elem| self.witness(*elem))
+            .collect::<Result<Vec<_>, CircuitError>>()?
+            .into_iter();
+        let sum_val = elem_iter.next().expect("cannot sum empty slice");
+        let sum_val = elem_iter.fold(sum_val, |acc, val| acc + val);
+        let sum = self.create_variable(sum_val)?;
+
+        // Pad with zeros and pack into as few gates as possibly by adding on all input
+        // wires
+        let mut padded: Vec<Variable> = elems.to_owned();
+        let gate_capacity = GATE_WIDTH - 1;
+        let padded_len = next_multiple(elems.len() - 1, gate_capacity)? + 1;
+        padded.resize(padded_len, self.zero());
+
+        // Construct a series of gates in which the output wire of the `n`th gate
+        // is the accumulation of 3 * n input wires in the sequence
+        let coeffs = [C::ScalarField::one(); GATE_WIDTH];
+        let mut accum = padded[0];
+        for i in 1..padded_len / gate_capacity {
+            accum = self.lc(
+                &[
+                    accum,
+                    padded[gate_capacity * i - 2],
+                    padded[gate_capacity * i - 1],
+                    padded[gate_capacity * i],
+                ],
+                &coeffs,
+            )?;
+        }
+
+        // Final round
+        let wires = [
+            accum,
+            padded[padded_len - 3],
+            padded[padded_len - 2],
+            padded[padded_len - 1],
+            sum,
+        ];
+        self.lc_gate(&wires, &coeffs)?;
+
+        Ok(sum)
+    }
+
+    /// Constrain variable `y` to the addition of `a` and `c`, where `c` is a
+    /// constant value
+    ///
+    /// Return an error if the input variables are invalid
+    fn add_constant_gate(
+        &mut self,
+        x: Variable,
+        c: C::ScalarField,
+        y: Variable,
+    ) -> Result<(), CircuitError> {
+        self.check_var_bound(x)?;
+        self.check_var_bound(y)?;
+
+        let wire_vars = &[x, self.one(), 0, 0, y];
+        self.insert_gate(wire_vars, Box::new(ConstantAdditionGate(c)))?;
+        Ok(())
+    }
+
+    /// Create a variable representing an addition with a constant value
+    ///
+    /// Return an error if the input variable is invalid
+    fn add_constant(
+        &mut self,
+        input_var: Variable,
+        elem: &C::ScalarField,
+    ) -> Result<Variable, CircuitError> {
+        self.check_var_bound(input_var)?;
+
+        let input_val = self.witness(input_var).unwrap();
+        let output_val = scalar!(*elem) + input_val;
+        let output_var = self.create_variable(output_val).unwrap();
+
+        self.add_constant_gate(input_var, *elem, output_var)?;
+
+        Ok(output_var)
+    }
+
+    /// Constrain variable `y` to the product of `a` and `c`, where `c` is a
+    /// constant value
+    ///
+    /// Return an error if the input variables are invalid.
+    fn mul_constant_gate(
+        &mut self,
+        x: Variable,
+        c: C::ScalarField,
+        y: Variable,
+    ) -> Result<(), CircuitError> {
+        self.check_var_bound(x)?;
+        self.check_var_bound(y)?;
+
+        let wire_vars = &[x, 0, 0, 0, y];
+        self.insert_gate(wire_vars, Box::new(ConstantMultiplicationGate(c)))?;
+        Ok(())
+    }
+
+    /// Create a variable representing a multiplication with a constant value
+    ///
+    /// Return an error if the input variable is invalid
+    fn mul_constant(
+        &mut self,
+        input_var: Variable,
+        elem: &C::ScalarField,
+    ) -> Result<Variable, CircuitError> {
+        self.check_var_bound(input_var)?;
+
+        let input_val = self.witness(input_var).unwrap();
+        let output_val = scalar!(*elem) * input_val;
+        let output_var = self.create_variable(output_val).unwrap();
+
+        self.mul_constant_gate(input_var, *elem, output_var)?;
+
+        Ok(output_var)
     }
 }
