+++ plonk/src/proof_system/snark.rs
@@ -40,7 +40,6 @@ use jf_primitives::{
 use jf_utils::par_utils::parallelizable_slice_iter;
 use mpc_relation::{
     constants::compute_coset_representatives, gadgets::ecc::SWToTEConParam, Arithmetization,
-    ConstraintSystem,
 };
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
